Moved Sanitize() method into Config.ts file to stop Webpack compilation issue
